cronjob-scripts/requirements.txt
amadeus
python-dotenv
mysql-connector-python